To version 0.5
To version 0.5

+++ src/org/hashuniforme/hash/funciones/FuncionHash.java
@@ -14,5 +14,5 @@ public interface FuncionHash {
 	 * @param o
 	 * @return
 	 */
-	public int getHash( String o);
+	public long getHash( String o);
 }
